requirements.txt
numpy==1.24.2
Pillow==9.5.0
pytesseract==0.3.10
opencv-python == 4.7.0.72